Set request timeout to 3 seconds
Since this is designed to be used for testing, most test cases
should be configured to use mocks, stubs, fakes, etc. and be
quick, and if something is wrong it should fail fast.
But if 3 seconds is not enough for you can always set the
client's timeout property yourself with the object you get
back from testclient.NewClient().

Maybe we should even set timeout to 1 second? After some
more field experience we'll see if we can.